requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.3
emoji==1.5.0
fastapi==0.70.1
h11==0.12.0
httptools==0.2.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.5
psycopg2==2.9.1
pydantic==1.8.2
pyspeedtest==1.2.7
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
typing_extensions==4.0.1
ujson==4.3.0
unittest-prettify==1.0.post4
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
